[BZ1652]: H263 video couldn't seek
root cause: the bug is caused by a special clips that uses P frame as key frame.

solution: All of the macroblocks of the key-P frame is intra coded and hence though it is marked as inter
frame, it could be independently decoded.

Remove the reference frame check for P frame within MIX so that the key frame could pass decoding process
when seeking.

Change-Id: I3a007732779433de83ba207f3a1ca9baf53532b8
Signed-off-by: Shuo Liu <shuo.liu@intel.com>