libmix: correct the method calucalte timestamp for mpeg4 decode.
BZ: 3035

Some mpeg4 clips freeze when playbacking, the timestamp is calculated uncorrectly, and framework will keep waiting to render the next frame
according to the wrong timestamp. When sometimes calculating the timestamp for a new frame, nanosecond is take as the basic unit, actually it should be macrosecond.

Change nanosecond to macrosecond as the basic unit for calcuating timestamp

Change-Id: I5bf57793900e0f60ceb849338b7929992297ea9f
Signed-off-by: xli111 <xiaowei.a.li@intel.com>