libmix: decoder: check driver render status before acquiring output surfaces
BZ: 25331

The video decoder checks for the rendering status before acquiring

Normal renderDone flag is used to track the surface status from OMX client
point of view.
There are cases where Android rendering pipeline thinks a frame is free
to be reused but that's not the case. The video driver knows better,
that's why we check with it. One of those cases are Widi and HDMI extended
video modes

Change-Id: I89db1ed8ff6244bffe3bec3bcc6867937bd920de
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/37262
Reviewed-by: Qiu, Junhai <junhai.qiu@intel.com>
Reviewed-by: Tao, Tao Q <tao.q.tao@intel.com>
Reviewed-by: Chen, Weian <weian.chen@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>