Enable Video Editor NV12 Processing on JB (BSP)
BZ: 51587

In video editor's original design, only IYUV420 processing is
supported. But the video format of Intel HW encoder and decoder
is NV12, we have to convert NV12 to IYUV420 on decoder side,
and IYUV420 to NV12 on encoder side. This is one bottleneck
for video editor, which slows down the video export process.

In this patch, we add NV12 support in video editor codes, which
eliminate the color conversion process. This significantly
accelerate the video export.

Change-Id: Iabd328de127b3a3b7224d6e6614666a00ec29df6
Signed-off-by: wfeng6 <wei.feng@intel.com>
Reviewed-on: http://android.intel.com:8080/62018
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>