Exit sharedbuffer mode when failing to create HW encoder
BZ: 58022

For the MFLD and CTP, there is a hardware limitation that we can
only encode one video stream at one time. When the encoder HW
is occupied by other apps, such as camera, video editor can't
create the HW encoder successfully. In this situation, we must
exit the sharedbuffer mode.

Change-Id: I4916621c2bc1fc55d40c1fcac75e443127d7d0af
Signed-off-by: wfeng6 <wei.feng@intel.com>
Reviewed-on: http://android.intel.com:8080/66989
Reviewed-by: Tang, Richard <richard.tang@intel.com>
Reviewed-by: Zhao, Leo <leo.zhao@intel.com>
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Jiang, Fei <fei.jiang@intel.com>
Reviewed-by: Wang, Yi A <yi.a.wang@intel.com>
Reviewed-by: Gu, Wangyi <wangyi.gu@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>