[PORT FROM R4.1]fix the messy issue when play the exported video againmainline BZ: 72393
BZ: 72393

root cause:
in the mix parser, the array size of frame store(fs) in struct h264_DecodedPictureBuffer
is 18, but will receive an index 31 to rewrite the entry value, unfortunately, this
behavior will overwrite the member of first_mb_in_slice in h264_Slice_Header_t for struct
h264_DecodedPictureBuffer and h264_Slice_Header_t are members of in the same struct of
h264_Info. The patch add an parameter check to fix the issue.

Change-Id: I6fba249a38673cf9c7af9ea9dc71063f470e01f7
Orig-Change-Id: I784fb6c76693b11a8c14d561755f7f1a92efb946
Signed-off-by: Gu, Wangyi <wangyi.gu@intel.com>
Signed-off-by: hding3 <haitao.ding@intel.com>
Reviewed-on: http://android.intel.com:8080/83610
Reviewed-by: Feng, Wei <wei.feng@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>