libmix: wait for the decode buffer ready if the surface is not composed by surfaceFlinger
BZ: 92956

if the surfaceTexture is not composed by surfaceFlinger, for example, video edit preview mode and HTML5
inline mode. the waiting for decode buffer ready should be done in decode thread to avoid tearing or color line
in such use case, because the app will not pass the NV12 buffer to HWC, and the NV12 buffer will be disposed by openGL
and convert to RGB in application.

Change-Id: I0a7f60b93f5ca1cbc2cf5cdcaf34f6044712a548
Signed-off-by: ywan171 <yi.a.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/100387
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>