Fix one VC1 clip playback issue
BZ: 113609

Sometimes, the offset calculated in AsfSimpleIndexParser::seek
is larger than the predefined threshold. In this situation,
the targetTime will be 0. This is the root cause of seeking
to the end but the progress bar goes to the beginning.

In this fix, we don't simply returns an error code. Instead,
the maximum value is given to the offset.

Also in this patch, the useless framework folder is removed.

Change-Id: I89ba770fc5ac0767b319845a9d9ef23fe74c3a6c
Signed-off-by: wfeng6 <wei.feng@intel.com>
Reviewed-on: http://android.intel.com:8080/113487
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>