Fix the widevine playback corruption problem on BYT
BZ: 140808

In current LibMIX implementation, there is one defect to handle
the MMCO commands. When the framenum of the slice is 0, the short
term reference frames will NOT be un-marked to be "un-referenced".
This makes the DPB management run into the error state and leads
to the corrupted decode frames.

Change-Id: I0907d5b357049d792318ec46cdc441c641257ae2
Signed-off-by: wfeng6 <wei.feng@intel.com>