Fix green bar issue on some videos
updateFormatInfo in AVC decoder defines new size of video by setting width and
height to a multiple of 16. Video size is really changed only if the aligned
width or height is different. handleNewSequence should match this behavior and
use aligned values of surfaceWidth and surfaceHeight which are defined in a
real video size change.

Change-Id: I62710239c7de5e4bba5e0efb5672b591096a6fdb